fixing errors, making new ones
